requirements.txt
absl-py==0.9.0
astor==0.8.1
astroid==2.4.1
astunparse==1.6.3
cachetools==4.1.0
certifi==2020.4.5.1
chardet==3.0.4
cycler==0.10.0
future==0.18.2
gast==0.3.3
grpcio==1.30.0
h5py==2.10.0
idna==2.9
importlib-metadata==1.6.0
isort==4.3.21
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.2
kiwisolver==1.2.0
labelImg==1.8.3
lazy-object-proxy==1.4.3
lxml==4.5.2
Markdown==3.2.2
matplotlib==3.2.1
mccabe==0.6.1
mock==4.0.2
numpy==1.18.4
oauthlib==3.1.0
opencv-python==4.2.0.34
opt-einsum==3.2.1
Pillow==7.1.2
protobuf==3.12.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyparsing==2.4.7
PyQt5==5.15.0
PyQt5-sip==12.8.0
python-dateutil==2.8.1
requests==2.23.0
requests-oauthlib==1.3.0
rsa==4.0
scipy==1.4.1
six==1.14.0
tensorboard==1.13.1
tensorflow==1.13.2
tensorflow-estimator==1.13.0
termcolor==1.1.0
terminaltables==3.1.0
toml==0.10.1
torch==1.5.0
torchvision==0.6.0
tqdm==4.46.0
typed-ast==1.4.1
urllib3==1.25.9
Werkzeug==1.0.1
wrapt==1.12.1
zipp==3.1.0
paho-mqtt==1.6.1
